CosmoParticles

Documentation for CosmoParticles.

CosmoParticles.AbstractParticles

CosmoParticles.Particles

CosmoParticles.Particles

CosmoParticles.get_props

CosmoParticles.particle_name

CosmoParticles.particle_number

CosmoParticles.show_properties

CosmoParticles.AbstractParticles — Type

AbstractParticles

Abstract supertype for storing particle data efficiently in arrays.

Particles generally have IDs, positions, and velocities as properties. The properties are expected to be stored in a Dict{Symbol,Any} as vectors of length $N$, matrices of size $m×N$, or as scalars (when the property is equal for all particles).

The inbuilt functionality of AbstractParticles includes accessing and setting properties via the following syntax:

p.id = [1, 2, 3]
p[:id] === p.id

If the struct has additional fields, these can also be accessed by p.field, but not by p[:field]. The latter syntax can only be used to access the particle properties.

Property keys

:id: Vector of IDs

:pos: $2×N$ or $3×N$ Matrix with positions

:vel: $2×N$ or $3×N$ Matrix with velocities

Methods

The methods Base.keys, Base.values, and Base.haskey are forwarded to the property Dict.

Concrete types of AbstractParticles should have the following methods implemented (also see the implementation of Particles):

Base.copy: returns new object containing a copy of the Dict

Base.:(==)

CosmoParticles.particle_name: returns the name of the struct to be printed via Base.show

Base.propertynames: implement this if there are additional struct fields

Base.show(io, mime, p)

CosmoParticles.Particles — Type

struct Particles <: AbstractParticles
    type::Symbol
    props::Dict{Symbol,Any}
end

Particles of a certain type (typically something like :dm or :gas in the cosmological context) with their properties.

CosmoParticles.Particles — Method

Particles(type[, pairs::Pair...])

Create a Particles object with the given type and pairs of Symbol to the property values. These are passed to the underlying Dict. If no pairs are passed to the method, an empty Dict is created.

CosmoParticles.get_props — Method

CosmoParticles.get_props(p::AbstractParticles)

Return the property Dict belonging to the particles.

This returns p.props by default if not overridden.

CosmoParticles.particle_name — Method

CosmoParticles.particle_name(::AbstractParticles)

Returns the name of the particle type, which is used when printing an object of this type via Base.show.

CosmoParticles.particle_number — Method

CosmoParticles.particle_number(p::AbstractParticles)

Returns the number of particles by determining the length of one of the property arrays.

CosmoParticles.show_properties — Method

CosmoParticles.show_properties(io::IO, mime, p::AbstractParticles)

Prints the number of particles, the name of the type, and the property names.

Should be used internally when overriding Base.show for concrete implementations of AbstractParticles.