Geometry

To filter particles in a particular volume, such as a cube, a sphere, or a cylinder, geometries can be used. This package provides a variety of different geometries for multiple dimensions, including the following:

Hyperrectangle

3D Rectangular Cuboid

2D Rectangle

Hypersphere

3D Sphere

2D Circle

Cylinder (standing cylinder aligned with the z axis and arbitrary orientation)

CosmoParticles.AbstractCosmoGeometry — Type

abstract type AbstractCosmoGeometry end

Abstract type for (multi-dimensional) geometry volumes, particularly for filtering with filter.

Any subtypes of AbstractCosmoGeometry have to implement the following methods:

CosmoParticles.geometry_enclosing_corners

CosmoParticles.geometry_enclosing_center

CosmoParticles.mask_in

CosmoParticles.geometry_enclosing_corners — Function

geometry_enclosing_corners(geo::AbstractCosmoGeometry)

Return the lower left and upper right corners of the enclosing box of the geometry as a tuple of vectors.

The enclosing box is not necessarily the tightest fitting box.

This is not exported.

CosmoParticles.geometry_enclosing_center — Function

geometry_enclosing_center(geo::AbstractCosmoGeometry)

Return the center of the enclosing box of the geometry as a vector.

This is not exported.

CosmoParticles.mask_in — Function

mask_in(pos::AbstractMatrix{<:Number}, geo::AbstractCosmoGeometry)

Return the BitArray mask of the positions ($\mathrm{dims} × N$) located within the geometry.

This is not exported.

CosmoParticles.CosmoHyperrectangle — Type

struct CosmoHyperrectangle{T,N} <: AbstractCosmoGeometry where {T<:Number}
    lowerleft::Vector{T}
    upperright::Vector{T}
end

Hyperrectangle aligned with the coordinate system axes given by its lower left and upper right corners.

The dimensions of space are given by N.

CosmoParticles.CosmoCuboid — Type

CosmoCuboid{T} = CosmoHyperrectangle{T,3}

Alias for a 3D CosmoHyperrectangle.

CosmoParticles.CosmoRectangle — Type

CosmoRectangle{T} = CosmoHyperrectangle{T,2}

Alias for a 2D CosmoHyperrectangle.

CosmoParticles.CosmoHypercube — Function

CosmoHypercube(center::AbstractVector{<:Number}, radius::Number)

Return a cubic CosmoCuboid centered around center, with equal sidelengths.

CosmoParticles.CosmoCube — Function

CosmoCube(center::AbstractVector{<:Number}, radius::Number)

Return a cubic CosmoCuboid centered around center, with equal sidelengths.

CosmoParticles.CosmoSquare — Function

CosmoSquare(center::AbstractVector{<:Number}, radius::Number)

Return a square CosmoRectangle centered around center, with equal sidelengths.

CosmoParticles.CosmoHypersphere — Type

struct CosmoHypersphere{T,N} <: AbstractCosmoGeometry where {T<:Number}
    center::Vector{T}
    radius::T
end

Hypersphere given by its center and radius.

If different types are passed to the constructor, the types will be promoted without throwing an error.

CosmoParticles.CosmoSphere — Type

CosmoSphere{T} = CosmoHypersphere{T,3}

Alias for a 2D CosmoHypersphere.

CosmoParticles.CosmoCircle — Type

CosmoCircle{T} = CosmoHypersphere{T,2}

Alias for a 2D CosmoHypersphere.

CosmoParticles.CosmoCylinder — Type

struct CosmoCylinder{T} <: AbstractCosmoGeometry where {T<:Number}
    startpos::Vector{T}
    endpos::Vector{T}
    radius::T
end

Cylinder given by its end points startpos and endpos and radius.

If different types are passed to the constructor, the types will be promoted without throwing an error.

CosmoParticles.CosmoStandingCylinder — Type

struct CosmoStandingCylinder{T} <: AbstractCosmoGeometry where {T<:Number}
    center::Vector{T}
    height::T
    radius::T
end

Standing cylinder given by its center, height, and radius.

Standing means that the cylinder is oriented such that its axis is aligned with the z axis. If different types are passed to the constructor, the types will be promoted without throwing an error.