Particle Collections

A set of different particles are represented by subtypes of AbstractParticleCollection. These can have a cosmological redshift attributed to them. In the context of cosmological simulations, this applies to collections of DM, gas, and star particles.

For an example implementation of a concrete subtype of AbstractParticleCollection, see the source codes of ParticleCollection or RedshiftParticleCollection.

CosmoParticles.AbstractParticleCollection — Type

abstract type AbstractParticleCollection{AP} end

Abstract supertype for storing data of multiple particle types.

The particles are expected to be stored in a Dict{Symbol,<:AbstractParticles}.

The inbuilt functionality of AbstractParticleCollection includes accessing and setting properties via the following syntax:

P<:AbstractParticles
pc::<:AbstractParticleCollection{P}
p::<:P
pc.dm = p
pc[:dm] === p

If the struct has additional fields, these can also be accessed by p.field, but not by p[:field]. The latter syntax can only be used to access the particles.

Methods

The methods Base.keys, Base.values, Base.haskey, Base.empty, Base.empty!, and Base.isempty are forwarded to the property Dict.

Concrete types of AbstractParticleCollection should have the following methods implemented (also see the implementation of ParticleCollection):

Base.copy: returns new object containing a copy of the Dict

Base.empty: returns an identical object, but with an empty Dict

Base.:(==)

redshift: implement this if the redshift of the particle collection is non-zero

Base.propertynames: implement this if there are additional struct fields

Base.show(io, mime, pc)

CosmoParticles.ParticleCollection — Type

struct ParticleCollection{AP<:AbstractParticles} <: AbstractParticleCollection{AP}
    particles::Dict{Symbol,AP}
end

Basic particle collection at redshift 0.

CosmoParticles.ParticleCollection — Method

ParticleCollection(::Type{AP}=Particles) where {AP<:AbstractParticles}

Create an empty particle collection of the passed type.

By default, the particle collection is created for the [Particles] type.

CosmoParticles.ParticleCollection — Method

ParticleCollection(pairs::Pair{Symbol,AP}...) where {AP<:AbstractParticles}

Create a particle collection from pairs of Symbols and particles.

Example

ParticleCollection(:dm => Particles(:dm), :gas => Particles(:gas))

CosmoParticles.ParticleCollection — Method

ParticleCollection(p::Particles...)

Create a particle collection from multiple Particles.

The Dict keys are taken from the Particles type.

Example

ParticleCollection(Particles(:dm), Particles(:gas))

CosmoParticles.RedshiftParticleCollection — Type

struct RedshiftParticleCollection{AP<:AbstractParticles,T<:Real} <: AbstractParticleCollection{AP}
    z::T
    particles::Dict{Symbol,AP}
end

Particle collection at a given redshift.

CosmoParticles.RedshiftParticleCollection — Method

RedshiftParticleCollection([::Type{AP}=Particles,] z::T) where {AP<:AbstractParticles,T<:Real}

Create an empty particle collection of the passed type at redshift z.

By default, the particle collection is created for the [Particles] type.

CosmoParticles.RedshiftParticleCollection — Method

RedshiftParticleCollection(z::T, pairs::Pair{Symbol,AP}...) where {AP<:AbstractParticles,T<:Real}

Create a particle collection at redshift z from pairs of Symbols and particles.

Example

RedshiftParticleCollection(0.5, :dm => Particles(:dm), :gas => Particles(:gas))

CosmoParticles.RedshiftParticleCollection — Method

RedshiftParticleCollection(z::Real, p::Particles...)

Create a particle collection at redshift z from multiple Particles.

The Dict keys are taken from the Particles type.

Example

RedshiftParticleCollection(0.5, Particles(:dm), Particles(:gas))

CosmoParticles.redshift — Function

redshift(pc::AbstractParticleCollection) = 0

Returns the redshift at which the particles are located.

By default, 0 is returned. This method should be overridden for concrete types of AbstractParticleCollection for which the redshift differs.

CosmoParticles.get_particles — Function

get_particles(pc::AbstractParticleCollection)

Return the particles Dict belonging to the particle collection.

This returns pc.particles by default if not overridden.

This is not exported.

CosmoParticles.show_properties — Method

CosmoParticles.show_properties(io::IO, mime, p::AbstractParticleCollection)

Prints the properties of the particles in the collection.

Should be used internally when overriding Base.show for concrete implementations of AbstractParticleCollection.

This is not exported.