Particles

A set of objects with different properties are represented by subtypes of AbstractParticles. While not absolutely necessary, these kind of objects generally have IDs, positions, and velocity as properties. In the context of cosmological simulations, this applies to particle and galaxy data.

For an example implementation of a concrete subtype of AbstractParticles, see the source code of Particles.

Missing docstring.

Missing docstring for CosmoParticles.AbstractParticles. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CosmoParticles.Particles. Check Documenter's build log for details.

CosmoParticles.get_props — Function

CosmoParticles.get_props(p::AbstractParticles)

Return the property Dict belonging to the particles.

This returns p.props by default if not overridden.

This is not exported.

CosmoParticles.particle_name — Function

CosmoParticles.particle_name(::AbstractParticles)

Returns the name of the particle type, which is used when printing an object of this type via Base.show.

This is not exported.

CosmoParticles.particle_number — Function

CosmoParticles.particle_number(p::AbstractParticles)

Returns the number of particles by determining the length of one of the property arrays.

This is not exported.

CosmoParticles.show_properties — Function

CosmoParticles.show_properties(io::IO, mime, p::AbstractParticles)

Prints the number of particles, the name of the type, and the property names.

Should be used internally when overriding Base.show for concrete implementations of AbstractParticles.

This is not exported.