CosmoParticles

Documentation for CosmoParticles.

CosmoParticles.AbstractParticles

CosmoParticles.Particles

CosmoParticles.Particles

Base.empty!

Base.filter

Base.filter

Base.filter!

Base.filter!

Base.sort

Base.sort!

CosmoParticles.applyind

CosmoParticles.applyind

CosmoParticles.applyind!

CosmoParticles.findall_in

CosmoParticles.findall_in_sorted

CosmoParticles.get_props

CosmoParticles.matrix_rotate

CosmoParticles.matrix_rotate!

CosmoParticles.particle_name

CosmoParticles.particle_number

CosmoParticles.rotate

CosmoParticles.show_properties

CosmoParticles.translate

CosmoParticles.translate!

LinearAlgebra.rotate!

CosmoParticles.AbstractParticles — Type

AbstractParticles

Abstract supertype for storing particle data efficiently in arrays.

Particles generally have IDs, positions, and velocities as properties. The properties are expected to be stored in a Dict{Symbol,Any} as vectors of length $N$, matrices of size $m×N$, or as scalars (when the property is equal for all particles).

The inbuilt functionality of AbstractParticles includes accessing and setting properties via the following syntax:

p.id = [1, 2, 3]
p[:id] === p.id

If the struct has additional fields, these can also be accessed by p.field, but not by p[:field]. The latter syntax can only be used to access the particle properties.

Property keys

:id: Vector of IDs

:pos: $2×N$ or $3×N$ Matrix with positions

:vel: $2×N$ or $3×N$ Matrix with velocities

Methods

The methods Base.keys, Base.values, and Base.haskey are forwarded to the property Dict.

Concrete types of AbstractParticles should have the following methods implemented (also see the implementation of Particles):

Base.copy: returns new object containing a copy of the Dict

Base.empty: returns an identical object, but with an empty Dict

Base.:(==)

CosmoParticles.particle_name: returns the name of the struct to be printed via Base.show

Base.propertynames: implement this if there are additional struct fields

Base.show(io, mime, p)

CosmoParticles.Particles — Type

struct Particles <: AbstractParticles
    type::Symbol
    props::Dict{Symbol,Any}
end

Particles of a certain type (typically something like :dm or :gas in the cosmological context) with their properties.

CosmoParticles.Particles — Method

Particles(type[, pairs::Pair...])

Create a Particles object with the given type and pairs of Symbol to the property values. These are passed to the underlying Dict. If no pairs are passed to the method, an empty Dict is created.

Base.empty! — Method

Base.empty!(p::AbstractParticles)

Remove all elements from the property Dict.

Base.filter! — Method

Base.filter!(p::AbstractParticles, ids)

Filter the particles in-place by keeping only those with the given IDs.

Base.filter! — Method

Base.filter!(f, p::AbstractParticles)

Filter the particles in-place by a mask returned by the function f.

The function f takes the particles as argument and has to either return a BitArray mask with a length of the number of particles or an array of indices.

Base.filter — Method

Base.filter(p::AbstractParticles, ids; affect=keys(p))

Create new particles with them filtered by keeping only those with the given IDs.

If the keyword argument affect is a non-empty tuple of Symbols, only those properties are filtered and added to the newly created particles object.

Base.filter — Method

Base.filter(f, p::AbstractParticles; affect=keys(p))

Create new particles with them filtered by a mask returned by the function f.

The function f takes the particles as argument and has to either return a BitArray mask with a length of the number of particles or an array of indices. If the keyword argument affect is a non-empty tuple of Symbols, only those properties are filtered and added to the newly created particles object.

Base.sort! — Method

Base.sort!(p::AbstractParticles, prop::Symbol; kwargs...)

Sort the particles in-place according to the specified property.

All property arrays of the particles are rearranged in-place according to the property being sorted. The keyword arguments are passed on to Base.sortperm.

The sorting algorithm for unitful properties may also be RadixSort from SortingAlgorithms.jl.

Base.sort — Method

Base.sort(p::AbstractParticles, prop::Symbol; affect=keys(p), kwargs...)

Create new particles with the particles sorted according to the specified property.

All property arrays of the particles are rearranged according to the property being sorted. If the keyword argument affect is a non-empty tuple of Symbols, only those properties are rearranged and added to the newly created particles object (besides the originally sorted property). The other keyword arguments are passed on to Base.sortperm.

The sorting algorithm for unitful properties may also be RadixSort from SortingAlgorithms.jl.

CosmoParticles.applyind! — Method

CosmoParticles.applyind!(p::AbstractParticles, ind::AbstractVector)

In-place application of indices or mask to all particle properties.

This is not exported.

CosmoParticles.applyind — Method

CosmoParticles.applyind(p::AbstractParticles, ind::AbstractVector; affect=keys(p))

Create new particles with the given indices or mask applied to all particle properties.

This can also be called by the simple syntax p[ind]. If the keyword argument affect is a non-empty tuple of Symbols, only those properties are indexed into and added to the newly created particles object.

This is not exported.

CosmoParticles.applyind — Method

CosmoParticles.applyind(a, ind::AbstractVector)

Apply indices or mask to a Number, Vector, or Matrix.

The following indexing is applied:

a::Number: a is returned directly.

a::Vector: a[ind] is returned.

a::Matrix: a[:, ind] is returned.

This is not exported.

CosmoParticles.findall_in — Method

CosmoParticles.findall_in(a::AbstractVector, set)

Return all indices of a that are in set.

If both a and set are sorted AbstractVectors, then the optimized findall_in_sorted is called. Otherwise, a Set is constructed from the Vector to perform the checks with in.

This is not exported.

CosmoParticles.findall_in_sorted — Method

CosmoParticles.findall_in_sorted(a::AbstractVector, set::AbstractVector)

Return all indices of a that are in set, where both a and set are assumed to be sorted.

This uses an optimized algorithm that is faster than creating a Set from set and performing checks with in.

This is not exported.

CosmoParticles.get_props — Method

CosmoParticles.get_props(p::AbstractParticles)

Return the property Dict belonging to the particles.

This returns p.props by default if not overridden.

CosmoParticles.matrix_rotate! — Method

CosmoParticles.matrix_rotate!(vals::AbstractMatrix{<:Number}, rotmat::AbstractMatrix{<:Real})

In-place version of CosmoParticles.matrix_rotate.

This is not exported.

CosmoParticles.matrix_rotate — Method

CosmoParticles.matrix_rotate(vals::AbstractMatrix{<:Number}, rotmat::AbstractMatrix{<:Real})

Returns vals rotated by the rotation matrix rotmat by regular matrix multiplication. Works for any dimensions and is optimized for unitful arrays.

This is not exported.

CosmoParticles.particle_name — Method

CosmoParticles.particle_name(::AbstractParticles)

Returns the name of the particle type, which is used when printing an object of this type via Base.show.

CosmoParticles.particle_number — Method

CosmoParticles.particle_number(p::AbstractParticles)

Returns the number of particles by determining the length of one of the property arrays.

CosmoParticles.rotate — Function

rotate(p::AbstractParticles, rotmat::AbstractMatrix{<:Real}, prop)
rotate(p::AbstractParticles, rotmat::AbstractMatrix{<:Real}, props=(:pos, :vel))

Rotates the specified properties props of the particles p by the rotation matrix rotmat.

Creates a copy of the particles with only new pointers to the rotated properties (by default position and velocity). This means that mutations to other properties will affect both the original particles as well as the rotated copied particles. The properties should be given as a single Symbol, or an array or tuple of Symbols. Only the existing quantities will be rotated (e.g., this method does not throw an error if velocities are not given for the particles).

CosmoParticles.show_properties — Method

CosmoParticles.show_properties(io::IO, mime, p::AbstractParticles)

Prints the number of particles, the name of the type, and the property names.

Should be used internally when overriding Base.show for concrete implementations of AbstractParticles.

CosmoParticles.translate — Function

translate(p::AbstractParticles, Δx, prop::Symbol=:pos)

Translates the specified property of the particles p by Δx.

Creates a copy of the particles with only new pointers to the translated property. Typically, Δx will be an AbstractVector, e.g., for positions or velocities.

CosmoParticles.translate! — Function

translate!(p::AbstractParticles, Δx, prop::Symbol=:pos)

In-place version of translate.

LinearAlgebra.rotate! — Function

LinearAlgebra.rotate!(p::AbstractParticles, rotmat::AbstractMatrix{<:Real}, props=(:pos, :vel))
LinearAlgebra.rotate!(p::AbstractParticles, rotmat::AbstractMatrix{<:Real}, prop)

In-place version of rotate.

This function is reexported.